Remove code which clears dataURL
In using `angular-signature` we found that clicking on the signature pad repeatedly would cause it to clear out.  I tested the `signature_pad` demo and I couldn't reproduce it, but I could reproduce it reliably in the `angular-signature` demo.  Doing some debugging it seemed to come from the `timeout` in `updateModel`.  When I removed the `isEmpty` check (as in this PR) we found that the problem went away and we didn't see any adverse effects.

So we were wondering: what was that code intended to do, and is it still needed?